refactor(bridge): extract incoming message dispatch
