+++ whatsrust/lib/main.go
@@ -1250,41 +1250,6 @@ func dispatchMessageActionEvent(action messageActionEvent) {
 	C.free(unsafe.Pointer(payload))
 }
 
-// dispatchIncomingMessage keeps action envelopes out of the ordinary message
-// path, whose insertion logic would otherwise treat an edit body as a base message.
-func dispatchIncomingMessage(
-	evt *events.Message,
-	dispatchAction func(messageActionEvent),
-	dispatchMessage func(types.MessageInfo, *waE2E.Message, bool),
-) {
-	dispatchIncomingMessageWithDecrypt(evt, dispatchAction, dispatchMessage, decryptSecretEncryptedMessage)
-}
-
-func dispatchIncomingMessageWithDecrypt(
-	evt *events.Message,
-	dispatchAction func(messageActionEvent),
-	dispatchMessage func(types.MessageInfo, *waE2E.Message, bool),
-	decrypt decryptSecretEncryptedMessageFunc,
-) {
-	if action, ok := messageActionEventFromSecretEncryptedMessage(evt, decrypt); ok {
-		dispatchAction(action)
-		return
-	}
-	if action, ok := messageActionEventFromIncomingMessage(evt); ok {
-		dispatchAction(action)
-		return
-	}
-	rawMessage := evt.RawMessage
-	if rawMessage == nil && evt.SourceWebMsg != nil {
-		rawMessage = evt.SourceWebMsg.GetMessage()
-	}
-	if message, ok := unavailableViewOnceMessage(rawMessage); ok {
-		dispatchMessage(evt.Info, message, false)
-		return
-	}
-	dispatchMessage(evt.Info, evt.Message, false)
-}
-
 func AddEventHandlers() {
 	client.AddEventHandler(func(rawEvt any) {
 		messageActionCensusDiagnostic(rawEvt)
